requirements.txt
psutil>=5.9.0
PyYAML>=6.0
jsonschema>=4.17.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0

